Feature/adding friendship requests

+++ FacebookClone/FacebookClone.Presentation/EndpointDefinitions/AlbumEndpointDefinition.cs
@@ -4,6 +4,7 @@ using FacebookClone.BLL.DTO.Image;
 using FacebookClone.BLL.Model;
 using FacebookClone.BLL.Services.Abstract;
 using FacebookClone.Presentation.Helpers;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 
@@ -13,9 +14,9 @@ namespace FacebookClone.Presentation.EndpointDefinitions
     {
         public static void DefineEndpoints(WebApplication app)
         {
-            app.MapGet("/albums", (IAlbumService albumService) => albumService.GetAll());
+            app.MapGet("/albums", [Authorize(Roles = "Admin", Policy = "RequireId")] (IAlbumService albumService) => albumService.GetAll());
 
-            app.MapPost("/albums", (HttpRequest request, IAlbumService albumService, IWebHostEnvironment environment) =>
+            app.MapPost("/albums", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (HttpRequest request, IAlbumService albumService, IWebHostEnvironment environment) =>
             {
                 if (request.Form.Files.Count == 0)
                 {
@@ -52,13 +53,16 @@ namespace FacebookClone.Presentation.EndpointDefinitions
                 return result;
             });
 
-            app.MapGet("/albums/{id}", (IAlbumService albumService, int id) => albumService.GetById(id));
+            app.MapGet("/albums/{id}", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (IAlbumService albumService, int id) => albumService.GetById(id));
 
-            app.MapGet("/albums/search/{userId}", (IAlbumService albumService, int userId, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => albumService.GetAllAlbumWithImagesByUserId(userId, pageSize, pageNumber));
+            app.MapGet("/albums/search/{userId}", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (IAlbumService albumService, int userId, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => albumService.GetAllAlbumWithImagesByUserId(userId, pageSize, pageNumber));
 
-            app.MapDelete("/albums/{id}", (IAlbumService albumService, int id) => albumService.Delete(id));
+            app.MapDelete("/albums/{id}", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (IAlbumService albumService, int id) => albumService.Delete(id));
 
-            app.MapPut("/albums", (AlbumDTO album, IAlbumService albumService) => albumService.Update(album));
+            app.MapPut("/albums", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (AlbumDTO album, IAlbumService albumService) => albumService.Update(album));
+
+            app.MapGet("/feed", [Authorize(Roles = "Admin,User", Policy = "RequireId")] (HttpContext context, IAlbumService albumService, [FromQuery(Name = "pageSize")] int pageSize, [FromQuery(Name = "pageNumber")] int pageNumber) => albumService.GetAllFriendsWithAlbumsWithImages(Convert.ToInt32(context.User.Claims.SingleOrDefault(e => e.Type == "id").Value), pageSize, pageNumber));
+
         }
     }
 }
